refactor(services): extract web scheduled task and context quality clients
Eleven methods across two independent contexts. Scheduled tasks bring
`scheduledTasks`, `nextScheduledTaskId` and `automaticArchivalSettings` with
them; all three are touched by no method outside this group, so they move
verbatim and stay unexported.

The context quality group holds no state of its own — it delegates to
`web-context-quality.ts` and returns a fixed manifest. `getContextEvidenceManifest`
calls `this.listContextEvidenceManifests`, so it declares `this: AgentService`.
